Link TimeTime action with glue #105
